Moved a bunch of test set up out of i18n_spec
Put it into i18n_spec_helper.rb
